moved batch_compute_resources into stacks folder and update it to be Stack instead of construct
